updated to emcee v3
