+++ docs/index.rst
@@ -1,11 +1,7 @@
 Welcome to beans's documentation!
 ======================================
 
-.. automodule:: beans.settle
-    :members:
-
-.. automodule:: beans.burstrain
-    :members:
+This software uses a Markov Chain Monte Carlo approach to match observations of an accreting neutron star in outburst with a simple ignition model (Settle, written by Andrew Cumming available in original form here - https://github.com/andrewcumming/settle). The code is all written in Python 3, except for settle which is a c++ code with a python wrapper. It makes use of Dan Foreman-Mackey's python implementation of MCMC, emcee, available here - https://github.com/dfm/emcee.
 
 .. toctree::
    :maxdepth: 2
